RAG-for-report/requirements.txt
# Core Environment
python-dotenv>=1.0.0
tqdm>=4.65.0


# Numerical & NLP Base
numpy<2.0
scikit-learn>=1.3.0
scipy>=1.10.0


# Database
mysql-connector-python>=8.0.0
pysqlite3-binary>=0.5.0


# NLP & Text Processing
konlpy>=0.6.0
JPype1>=1.4.0
keybert>=0.8.0
sentence-transformers>=2.2.0


# Vector DB & Retrieval
chromadb>=0.4.0
langchain>=0.1.20
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-ollama>=0.1.0


# LLM API
google-generativeai>=0.8.0
